fixed: code review comments
