--- src/chmock/filter.js
@@ -1,9 +1,0 @@
-import {filter} from '../utils/filter.js';
-
-const generateFilter = (points) => Object.entries(filter).map(
-  ([filterName, filterPoints]) => ({
-    name: filterName,
-    count: filterPoints(points).length,
-  }),
-);
-export default generateFilter;
