+++ src/main.js
@@ -1,5 +1,4 @@
 import {render} from './framework/render.js';
-import generateFilter from './chmock/filter.js';
 
 import FilterView from './view/filter-view.js';
 import TripPresenter from './presenter/trip-presenter.js';
@@ -14,10 +13,9 @@ const tripContainerElement = document.querySelector('.trip-events');
 const pointsModel = new PointsModel();
 const offersModel = new OffersModel();
 const destinationsModel = new DestinationsModel();
-const filters = generateFilter(pointsModel.points);
 
 const tripPresenter = new TripPresenter(tripContainerElement, pointsModel, offersModel, destinationsModel);
 
-render(new FilterView(filters), filterContainerElement);
+render(new FilterView(pointsModel.filteredPoints), filterContainerElement);
 
 tripPresenter.init();
