--- src/utils/filter.js
@@ -1,10 +1,0 @@
-import {FilterType} from '../constants.js';
-import {isPointInThePast, isPointInTheFuture} from './points.js';
-
-const filter = {
-  [FilterType.EVERYTHING]: (points) => points,
-  [FilterType.PAST]: (points) => points.filter((point) => isPointInThePast(point.dateTo)),
-  [FilterType.FUTURE]: (points) => points.filter((point) => isPointInTheFuture(point.dateFrom)),
-};
-
-export {filter};
